move runtime/golang/src to runtime/golang
