+++ serde-generate/runtime/golang/go.mod
@@ -1,0 +1,5 @@
+module github.com/facebookincubator/serde-reflection/serde-generate/runtime/golang
+
+go 1.14
+
+require github.com/stretchr/testify v1.6.1
