+++ serde-generate/runtime/golang/serde/interfaces.go
