--- serde-generate/runtime/golang/src/go.mod
@@ -1,5 +1,0 @@
-module github.com/facebookincubator/serde-reflection/serde-generate/runtime/golang/src
-
-go 1.14
-
-require github.com/stretchr/testify v1.6.1
